+++ src/test/merkleblock_tests.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2012-2017 The Bitcoin Core developers
-// Copyright (c) 2017 The Raven Core developers
+// Copyright (c) 2017-2019 The Raven Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -38,20 +38,20 @@ BOOST_FIXTURE_TEST_SUITE(merkleblock_tests, BasicTestingSetup)
         BOOST_CHECK_EQUAL(merkleBlock.header.GetHash().GetHex(), block.GetHash().GetHex());
 
         // vMatchedTxn is only used when bloom filter is specified.
-        BOOST_CHECK_EQUAL(merkleBlock.vMatchedTxn.size(), 0);
+        BOOST_CHECK_EQUAL(merkleBlock.vMatchedTxn.size(), (uint64_t)0);
 
         std::vector<uint256> vMatched;
         std::vector<unsigned int> vIndex;
 
         BOOST_CHECK_EQUAL(merkleBlock.txn.ExtractMatches(vMatched, vIndex).GetHex(), block.hashMerkleRoot.GetHex());
-        BOOST_CHECK_EQUAL(vMatched.size(), 2);
+        BOOST_CHECK_EQUAL(vMatched.size(), (uint64_t)2);
 
         // Ordered by occurrence in depth-first tree traversal.
         BOOST_CHECK_EQUAL(vMatched[0].ToString(), txhash2.ToString());
-        BOOST_CHECK_EQUAL(vIndex[0], 1);
+        BOOST_CHECK_EQUAL(vIndex[0], (uint64_t)1);
 
         BOOST_CHECK_EQUAL(vMatched[1].ToString(), txhash1.ToString());
-        BOOST_CHECK_EQUAL(vIndex[1], 8);
+        BOOST_CHECK_EQUAL(vIndex[1], (uint64_t)8);
     }
 
 
@@ -70,14 +70,14 @@ BOOST_FIXTURE_TEST_SUITE(merkleblock_tests, BasicTestingSetup)
         CMerkleBlock merkleBlock(block, txids2);
 
         BOOST_CHECK_EQUAL(merkleBlock.header.GetHash().GetHex(), block.GetHash().GetHex());
-        BOOST_CHECK_EQUAL(merkleBlock.vMatchedTxn.size(), 0);
+        BOOST_CHECK_EQUAL(merkleBlock.vMatchedTxn.size(), (uint64_t)0);
 
         std::vector<uint256> vMatched;
         std::vector<unsigned int> vIndex;
 
         BOOST_CHECK_EQUAL(merkleBlock.txn.ExtractMatches(vMatched, vIndex).GetHex(), block.hashMerkleRoot.GetHex());
-        BOOST_CHECK_EQUAL(vMatched.size(), 0);
-        BOOST_CHECK_EQUAL(vIndex.size(), 0);
+        BOOST_CHECK_EQUAL(vMatched.size(), (uint64_t)0);
+        BOOST_CHECK_EQUAL(vIndex.size(), (uint64_t)0);
     }
 
 BOOST_AUTO_TEST_SUITE_END()
